Update go-faiss and run go mod tidy

+++ internal/index/metadata_embedding.go
@@ -61,7 +61,7 @@ func (idx *MetadataIndex) Delete() {
 // Query queries the index with vec.
 // Returns the dataset IDs of the k nearest neighbors and the corresponding
 // cosine similarity, sorted by similarity.
-func (idx *MetadataIndex) Query(vec []float32, k int) ([]string, []float32, error) {
+func (idx *MetadataIndex) Query(vec []float32, k int64) ([]string, []float32, error) {
 	dist, ids, err := idx.idx.Search(vec, k)
 	if err != nil {
 		return nil, nil, err
